Safari-609.1.11.1/LayoutTests/js/dom/unhandled-promise-rejection-bindings-type-error-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: The PromiseRejectionEvent.promise getter can only be used on instances of PromiseRejectionEvent
Test rejected promises are returned from bindings and trigger unhandledrejection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS error.promise is promise
PASS error.reason instanceof TypeError is true
PASS successfullyParsed is true

TEST COMPLETE
